Change MDP time-sampling dim with forecast or analysis data
The mllam-data-prep datastore previously assumed that the time-sampling
dimension is always called `time`, but when working with forecast data the
time-sampling dimension is actually assumed to be called `analysis_time` rather
than `time`. This commit introduces a class property on the mdp datastore that
changes with the type of data being used.